// projects/website/src/app/interfaces/questions.ts
export interface Questions {
  question_id: string;
  question_english: string;
  correct_answer_english: string;
  incorrect_answer_1_english: string;
  incorrect_answer_2_english: string;
  incorrect_answer_3_english: string;
  question_french: string;
  correct_answer_french: string;
  incorrect_answer_1_french: string;
  incorrect_answer_2_french: string;
  incorrect_answer_3_french: string;
}